Cambio: Tipo de

+++ compartimiento.cpp
@@ -1,7 +1,7 @@
 #include "compartimiento.h"
 
 //Constructor
-Compartimiento::Compartimiento(double capacidad): capacidad_(capacidad), cantidadActual_(0.0) {}
+Compartimiento::Compartimiento(float capacidad): capacidad_(capacidad), cantidadActual_(0.0) {}
 
 //Getters
 double Compartimiento::getCapacidad() const {
@@ -11,6 +11,6 @@ double Compartimiento::getCantidadActual() const{
     return cantidadActual_;
 }
 //Setters
-void Compartimiento::setCantidadActual(double cantidad){
+void Compartimiento::setCantidadActual(float cantidad){
     cantidadActual_ = cantidad;
 }
